Wed, Jun 12, 2024, 2:04 PM -03:00

+++ pset4/lab4/volume.c
@@ -7,7 +7,7 @@
 // Number of bytes in .wav header
 const int HEADER_SIZE = 44;
 typedef uint8_t BYTE;
-typedef int8_t BYTES2;
+typedef int16_t BYTES2;
 int main(int argc, char *argv[])
 {
     // Check command-line arguments
